+++ backend/test/api/helper_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/kinvolk/nebraska/backend/pkg/api"
+	"github.com/flatcar/nebraska/backend/pkg/api"
 )
 
 // newDBForTest is a helper function that